# ai_stock_analyzer/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Stock Analyzer API
  description: API untuk mengambil data harga saham historis, indikator teknikal (VPA/Wyckoff), dan analisis/rekomendasi harian berbasis AI.
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /stocks/{ticker}/chart:
    get:
      tags:
        - Stocks
      summary: Dapatkan data harga riwayat (OHLCV) untuk chart
      parameters:
        - in: path
          name: ticker
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
  /stocks/{ticker}/sync:
    post:
      tags:
        - Stocks
      summary: Sinkronisasi data saham dari yfinance
      parameters:
        - in: path
          name: ticker
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
  /stocks/{ticker}/analysis/latest:
    get:
      tags:
        - Analysis
      summary: Dapatkan hasil analisis AI dan teknikal terbaru untuk sebuah saham
      parameters:
        - in: path
          name: ticker
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
  /stocks/watchlist:
    get:
      tags:
        - Stocks
      summary: Dapatkan daftar watchlist pengguna
      responses:
        '200':
          description: Success
    post:
      tags:
        - Stocks
      summary: Tambahkan saham ke watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ticker:
                  type: string
                notes:
                  type: string
      responses:
        '201':
          description: Created
  /stocks/watchlist/{ticker}:
    delete:
      tags:
        - Stocks
      summary: Hapus saham dari watchlist
      parameters:
        - in: path
          name: ticker
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
...